Move YouTube and Twitter webhooks into utils
